+++ client/client_test.go
@@ -9,88 +9,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/mattermost/rtcd/service/random"
-
 	"github.com/stretchr/testify/require"
 )
 
-func TestConfigParse(t *testing.T) {
-	t.Run("empty struct", func(t *testing.T) {
-		var cfg Config
-		err := cfg.Parse()
-		require.Error(t, err)
-		require.Equal(t, "invalid SiteURL value: should not be empty", err.Error())
-	})
-
-	t.Run("invalid SiteURL scheme", func(t *testing.T) {
-		cfg := Config{
-			SiteURL: "ws://host",
-		}
-		err := cfg.Parse()
-		require.Error(t, err)
-		require.Equal(t, "invalid SiteURL scheme \"ws\"", err.Error())
-	})
-
-	t.Run("spaces in SiteURL", func(t *testing.T) {
-		cfg := Config{
-			SiteURL:   " http://host  ",
-			AuthToken: random.NewID(),
-		}
-		err := cfg.Parse()
-		require.NoError(t, err)
-	})
-
-	t.Run("slashes in SiteURL", func(t *testing.T) {
-		cfg := Config{
-			SiteURL:   "http://host/subpath////",
-			AuthToken: random.NewID(),
-		}
-		err := cfg.Parse()
-		require.NoError(t, err)
-		require.Equal(t, "http://host/subpath", cfg.SiteURL)
-	})
-
-	t.Run("empty AuthToken", func(t *testing.T) {
-		cfg := Config{
-			SiteURL: "http://mm-url",
-		}
-		err := cfg.Parse()
-		require.Error(t, err)
-		require.Equal(t, "invalid AuthToken value: should not be empty", err.Error())
-	})
-
-	t.Run("wsURL", func(t *testing.T) {
-		cfg := Config{
-			SiteURL:   "https://mm-url:8065/",
-			AuthToken: random.NewID(),
-		}
-		err := cfg.Parse()
-		require.NoError(t, err)
-		require.Equal(t, "wss://mm-url:8065/api/v4/websocket", cfg.wsURL)
-
-		cfg.SiteURL = "http://mm-url//"
-		err = cfg.Parse()
-		require.NoError(t, err)
-		require.Equal(t, "ws://mm-url/api/v4/websocket", cfg.wsURL)
-
-		cfg.SiteURL = "http://mm-url/subpath/"
-		err = cfg.Parse()
-		require.NoError(t, err)
-		require.Equal(t, "ws://mm-url/subpath/api/v4/websocket", cfg.wsURL)
-	})
-
-	t.Run("valid", func(t *testing.T) {
-		cfg := Config{
-			SiteURL:   "https://mm-url:8065/",
-			AuthToken: random.NewID(),
-		}
-		err := cfg.Parse()
-		require.NoError(t, err)
-	})
-}
-
 func TestClientConnect(t *testing.T) {
-	th := setupTestHelper(t)
+	th := setupTestHelper(t, "")
 
 	connectCh := make(chan struct{})
 	th.userClient.On(WSConnectEvent, func() error {
@@ -110,7 +33,7 @@ func TestClientConnect(t *testing.T) {
 		return nil
 	})
 
-	err := th.userClient.Connect(random.NewID())
+	err := th.userClient.Connect()
 	require.NoError(t, err)
 
 	select {
@@ -136,15 +59,15 @@ func TestClientConnect(t *testing.T) {
 }
 
 func TestClientConsistency(t *testing.T) {
-	th := setupTestHelper(t)
+	th := setupTestHelper(t, "")
 
 	t.Run("double connect", func(t *testing.T) {
-		err := th.userClient.Connect(random.NewID())
+		err := th.userClient.Connect()
 		require.NoError(t, err)
 
 		require.Equal(t, clientStateInit, th.userClient.state)
 
-		err = th.userClient.Connect(random.NewID())
+		err = th.userClient.Connect()
 		require.EqualError(t, err, "ws client is already initialized")
 	})
 
@@ -159,13 +82,13 @@ func TestClientConsistency(t *testing.T) {
 	})
 
 	t.Run("reuse client", func(t *testing.T) {
-		err := th.userClient.Connect(random.NewID())
+		err := th.userClient.Connect()
 		require.EqualError(t, err, "ws client is already initialized")
 	})
 }
 
 func TestClientConcurrency(t *testing.T) {
-	th := setupTestHelper(t)
+	th := setupTestHelper(t, "")
 
 	connectCh := make(chan struct{})
 	closeCh := make(chan struct{})
@@ -197,7 +120,7 @@ func TestClientConcurrency(t *testing.T) {
 		go func() {
 			defer wg.Done()
 
-			err := th.userClient.Connect(random.NewID())
+			err := th.userClient.Connect()
 			if err != nil {
 				atomic.AddInt32(&connectErrors, 1)
 			}
